Added Archer class Added Archer First Skill (Focus) - chance to double attack damage per standard attack (PASSIVE) Added Archer Second Skill (StraightShot) - Fires an explosive arrow Fixed Trainer bug (Mobs would attack it and crash the server) Changed Weapon Shapes to switch between lib files. 0-99 Standard weapons. 100-199 Assassin Weapons. 200-299 Archer Weapons.

+++ Server/ItemInfoForm.cs
@@ -501,9 +501,9 @@ namespace Server
         {
             if (ActiveControl != sender) return;
 
-            sbyte temp;
+            short temp;
 
-            if (!sbyte.TryParse(ActiveControl.Text, out temp))
+            if (!short.TryParse(ActiveControl.Text, out temp))
             {
                 ActiveControl.BackColor = Color.Red;
                 return;
